+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Ability to select payment option when submitting jobs from the Python client.
 - Periodic repetition of EME subgrids via `num_reps` or `EMEPeriodicitySweep`.
 - Methods `EMEExplicitGrid.from_structures` and `EMECompositeGrid.from_structure_groups` to place EME cell boundaries at structure bounds.
+- 'ModeSimulation' now supports 'PermittivityMonitor'.
 
 ### Changed
 - Performance enhancement for adjoint gradient calculations by optimizing field interpolation.